Segregate TargetChain events
